backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.43.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
pydantic==2.5.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
